authentik initial config
we add postgres, redis, authentik itself, plus a whoami service. we set
up init scripts

+++ install.sh
@@ -9,6 +9,46 @@ source .env
 # for traefik acme certs
 mkdir -p ${STORAGE}/traefik/acme
 
+# for the cloudflare credentials
+if [ ! -f "${STORAGE}/traefik/cloudflare.env" ]
+then
+  echo "Please create ${STORAGE}/traefik/cloudflare.env with the following content:"
+  echo "CLOUDFLARE_DNS_API_TOKEN=..."
+  exit 1
+fi
+
+# mmrs data dir
+mkdir -p ${STORAGE}/mmrs/data
+chown 82:82 ${STORAGE}/mmrs/data
+
+# redis
+mkdir -p ${STORAGE}/redis/data
+
+# postgres
+mkdir -p ${STORAGE}/postgres/data
+if [ ! -f "${STORAGE}/postgres/env" ]
+then
+  echo "Please create ${STORAGE}/postgres/env with the following content:"
+  echo "PG_ADMIN_USER=..."
+  echo "PG_ADMIN_PASS=..."
+  exit 1
+fi
+
+# authentik config
+mkdir -p ${STORAGE}/authentik/media ${STORAGE}/authentik/templates
+if [ ! -f "${STORAGE}/authentik/env" ]
+then
+  echo "Please create ${STORAGE}/authentik/env with the following content:"
+  echo "AUTHENTIK_PG_USER=..."
+  echo "AUTHENTIK_PG_PASS=..."
+  echo "AUTHENTIK_PG_DB=..."
+  echo "AUTHENTIK_SECRET_KEY=..."
+
+  exit 1
+fi
+
+
+
 # install the systemd unit file
 cp ${REPO}/compose.service /etc/systemd/system/
 systemctl daemon-reload
